requirements.txt
aioredis==1.3.1
APScheduler==3.8.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
daphne==3.0.2
Django==3.1.7
django-environ==0.8.1
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
django-easy-pdf
et-xmlfile==1.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
openpyxl==3.0.7
pandas==1.3.4
paramiko==2.8.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
requests==2.25.1
scp==0.14.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.4
XlsxWriter==3.0.1
zope.interface==5.3.0
